+++ mpu6050/Makefile
@@ -3,7 +3,8 @@
 #
 ifneq ($(KERNELRELEASE),)
 
-obj-m := mpu6050.o
+obj-m += mpu6050.o
+mpu6050-objs := mpu6050_sysfs.o mpu6050_main.o
 
 else
